# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fuel Nutrition API",
    "version": "1.0.0",
    "description": "Nutrition tracking API for the Fuel AI coach. Logs food, tracks weight, manages goals, and provides coaching insights."
  },
  "servers": [
    {
      "url": "https://fuel-2j8v.onrender.com/api",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get today's dashboard",
        "description": "Returns daily energy/macro budget, logged meals, weight trend series, adaptive expenditure model, adherence, and goal progress.",
        "operationId": "getDashboard",
        "responses": {
          "200": {
            "description": "Dashboard data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/weight": {
      "get": {
        "tags": [
          "Weight"
        ],
        "summary": "Get weight trend data",
        "description": "Weight tracking view with trend series, smoothed stats, and headline metrics for a selected window.",
        "operationId": "getWeight",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                30,
                60,
                90,
                180
              ],
              "default": 90
            },
            "description": "Lookback window in days."
          }
        ],
        "responses": {
          "200": {
            "description": "Weight data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Weight"
        ],
        "summary": "Log a weigh-in",
        "description": "Record or overwrite a scale weight for a given date. Returns updated WeightData.",
        "operationId": "logWeight",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                30,
                60,
                90,
                180
              ],
              "default": 90
            },
            "description": "Lookback window for the response."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "weight"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "ISO date (YYYY-MM-DD)."
                  },
                  "weight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "maximum": 500,
                    "description": "Scale weight in kg."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Weigh-in recorded — returns updated weight data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightData"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date or weight out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/insights": {
      "get": {
        "tags": [
          "Insights"
        ],
        "summary": "Get aggregated insights",
        "description": "Weekly summaries, plateau detection, adherence stats, forward projection to goal, and human-readable insight cards.",
        "operationId": "getInsights",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GoalMode"
            },
            "description": "Goal mode."
          },
          {
            "name": "targetWeight",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 78,
              "minimum": 1,
              "maximum": 500
            },
            "description": "Target weight in kg."
          }
        ],
        "responses": {
          "200": {
            "description": "Insights data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/progress": {
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "Get progress data",
        "description": "Body measurements, lift history with PRs, cardio entries, step trend, and streak stats. All generated from deterministic sample data.",
        "operationId": "getProgress",
        "responses": {
          "200": {
            "description": "Progress data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/goal": {
      "get": {
        "tags": [
          "Goal"
        ],
        "summary": "Get current goal",
        "description": "Returns the active user goal configuration (mode, target weight, start weight, start date).",
        "operationId": "getGoal",
        "responses": {
          "200": {
            "description": "Goal configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoal"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Goal"
        ],
        "summary": "Set or update goal",
        "description": "Overwrite the active goal. Missing optional fields fall back to the existing values.",
        "operationId": "setGoal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mode",
                  "targetWeight"
                ],
                "properties": {
                  "mode": {
                    "$ref": "#/components/schemas/GoalMode"
                  },
                  "targetWeight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "description": "Target body weight in kg."
                  },
                  "startWeight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "description": "Starting body weight in kg (optional, falls back to existing)."
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date",
                    "description": "ISO start date (optional, falls back to existing)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goal updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid mode or targetWeight",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/coach": {
      "get": {
        "tags": [
          "Coach"
        ],
        "summary": "Get coach briefing",
        "description": "Deterministic AI coach briefing with prioritized recommendations, check-in summary, macro targets, and talking points — all grounded in the user's training data.",
        "operationId": "getCoachBriefing",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GoalMode"
            },
            "description": "Goal mode."
          },
          {
            "name": "targetWeight",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 78,
              "minimum": 1,
              "maximum": 500
            },
            "description": "Target weight in kg."
          },
          {
            "name": "currentTarget",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 20000
            },
            "description": "Current daily calorie target. Defaults to engine recommendation."
          }
        ],
        "responses": {
          "200": {
            "description": "Coach briefing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/reset": {
      "post": {
        "tags": [
          "Data"
        ],
        "summary": "Reset all data to defaults",
        "description": "Clears food log, weight history, and resets goal to default. Deletes persisted store file.",
        "operationId": "resetAllData",
        "responses": {
          "200": {
            "description": "All data reset",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/user": {
      "get": {
        "tags": ["User"],
        "summary": "Get current user info",
        "description": "Returns user info for the authenticated user. Requires X-User-Id header.",
        "operationId": "getUser",
        "parameters": [
          { "$ref": "#/components/parameters/UserIdHeader" }
        ],
        "responses": {
          "200": {
            "description": "User info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["userId", "createdAt"],
                  "properties": {
                    "userId": { "type": "string", "format": "uuid" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "UserIdHeader": {
        "name": "X-User-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "format": "uuid" },
        "description": "User identifier from POST /api/user/register. If omitted, uses the default shared user."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message."
          }
        }
      },
...
      "GoalMode": {
        "type": "string",
        "enum": [
          "fat-loss",
          "maintenance",
          "lean-bulk",
          "recomp"
        ],
        "description": "Nutrition goal mode."
      },
...
      "MacroTarget": {
        "type": "object",
        "required": [
          "target",
          "consumed",
          "remaining"
        ],
        "properties": {
          "target": {
            "type": "integer",
            "description": "Budgeted amount."
          },
          "consumed": {
            "type": "integer",
            "description": "Amount logged so far."
          },
          "remaining": {
            "type": "integer",
            "description": "Amount still available."
          }
        }
      },
      "DashboardMacros": {
        "type": "object",
        "required": [
          "protein",
          "carbs",
          "fat"
        ],
        "properties": {
          "protein": {
            "$ref": "#/components/schemas/MacroTarget"
          },
          "carbs": {
            "$ref": "#/components/schemas/MacroTarget"
          },
          "fat": {
            "$ref": "#/components/schemas/MacroTarget"
          }
        }
      },
      "Meal": {
        "type": "object",
        "required": [
          "id",
          "name",
          "calories",
          "protein",
          "carbs",
          "fat"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "calories": {
            "type": "integer"
          },
          "protein": {
            "type": "integer"
          },
          "carbs": {
            "type": "integer"
          },
          "fat": {
            "type": "integer"
          },
          "time": {
            "type": "string",
            "description": "Optional time label."
          }
        }
      },
      "WeightPoint": {
        "type": "object",
        "required": [
          "date",
          "scale",
          "trend"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "scale": {
            "type": "number",
            "nullable": true,
            "description": "Raw scale weight (null if no weigh-in that day)."
          },
          "trend": {
            "type": "number",
            "description": "Smoothed trend weight."
          }
        }
      },
      "EnergyModel": {
        "type": "object",
        "required": [
          "expenditureEstimate",
          "confidence",
          "trendDelta"
        ],
        "properties": {
          "expenditureEstimate": {
            "type": "number",
            "description": "Estimated TDEE in kcal."
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Model confidence (0–1)."
          },
          "trendDelta": {
            "type": "number",
            "description": "Expenditure change vs 2 weeks ago."
          }
        }
      },
      "GoalProgress": {
        "type": "object",
        "required": [
          "mode",
          "targetWeightDelta",
          "progress",
          "etaWeeks"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "targetWeightDelta": {
            "type": "number",
            "description": "Distance from target (kg, signed)."
          },
          "progress": {
            "type": "number",
            "description": "Fraction of goal achieved (0–1+)."
          },
          "etaWeeks": {
            "type": "number",
            "nullable": true,
            "description": "Estimated weeks to goal (null if diverging)."
          }
        }
      },
      "DashboardData": {
        "type": "object",
        "required": [
          "date",
          "calories",
          "macros",
          "meals",
          "weightSeries",
          "energy",
          "weeklyAdherence",
          "goal"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "calories": {
            "$ref": "#/components/schemas/MacroTarget",
            "description": "Calorie budget."
          },
          "macros": {
            "$ref": "#/components/schemas/DashboardMacros"
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            },
            "description": "Today's logged meals."
          },
          "weightSeries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeightPoint"
            },
            "description": "Trailing weight trend (oldest to newest)."
          },
          "energy": {
            "$ref": "#/components/schemas/EnergyModel"
          },
          "weeklyAdherence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Fraction of last 7 days on target."
          },
          "goal": {
            "$ref": "#/components/schemas/GoalProgress"
          },
          "notes": {
            "type": "string",
            "description": "Optional free-text note."
          }
        }
      },
      "WeightStats": {
        "type": "object",
        "required": [
          "currentTrend",
          "latestScale",
          "weeklyRate",
          "totalChange",
          "lowestTrend",
          "highestTrend",
          "entriesLogged",
          "windowDays"
        ],
        "properties": {
          "currentTrend": {
            "type": "number",
            "description": "Latest smoothed trend weight (kg)."
          },
          "latestScale": {
            "type": "number",
            "nullable": true,
            "description": "Most recent raw scale weight (kg)."
          },
          "weeklyRate": {
            "type": "number",
            "description": "Trend change over last 7 days (kg, signed)."
          },
          "totalChange": {
            "type": "number",
            "description": "Trend change across full window (kg, signed)."
          },
          "lowestTrend": {
            "type": "number",
            "description": "Lowest trend weight in window (kg)."
          },
          "highestTrend": {
            "type": "number",
            "description": "Highest trend weight in window (kg)."
          },
          "entriesLogged": {
            "type": "integer",
            "description": "Days with a raw weigh-in in window."
          },
          "windowDays": {
            "type": "integer",
            "description": "Total days in the window."
          }
        }
      },
      "WeightData": {
        "type": "object",
        "required": [
          "date",
          "range",
          "series",
          "stats",
          "unit"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date of the newest data point."
          },
          "range": {
            "type": "integer",
            "enum": [
              30,
              60,
              90,
              180
            ],
            "description": "Lookback window."
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeightPoint"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/WeightStats"
          },
          "unit": {
            "type": "string",
            "enum": [
              "kg",
              "lb"
            ],
            "description": "Weight unit."
          }
        }
      },
      "WeekSummary": {
        "type": "object",
        "required": [
          "weekStart",
          "weekEnd",
          "label",
          "avgIntake",
          "avgTrendWeight",
          "weeklyRate",
          "adherence",
          "daysLogged"
        ],
        "properties": {
          "weekStart": {
            "type": "string",
            "format": "date",
            "description": "Monday of the week."
          },
          "weekEnd": {
            "type": "string",
            "format": "date",
            "description": "Sunday of the week."
          },
          "label": {
            "type": "string",
            "description": "Human label (e.g. 'Jul 1 – Jul 7')."
          },
          "avgIntake": {
            "type": "number",
            "nullable": true,
            "description": "Mean intake (kcal) across days with logs."
          },
          "avgTrendWeight": {
            "type": "number",
            "description": "Mean trend weight (kg)."
          },
          "weeklyRate": {
            "type": "number",
            "nullable": true,
            "description": "Trend change from previous week (kg)."
          },
          "adherence": {
            "type": "number",
            "description": "Fraction of days logged that hit target."
          },
          "daysLogged": {
            "type": "integer"
          }
        }
      },
      "PlateauStatus": {
        "type": "object",
        "required": [
          "detected",
          "weeks",
          "message"
        ],
        "properties": {
          "detected": {
            "type": "boolean"
          },
          "weeks": {
            "type": "integer",
            "description": "Consecutive flat weeks."
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AdherenceSummary": {
        "type": "object",
        "required": [
          "overall",
          "loggingRate",
          "onTargetDays",
          "totalLoggedDays",
          "totalDays"
        ],
        "properties": {
          "overall": {
            "type": "number",
            "description": "Fraction of logged days within ±15% of target."
          },
          "loggingRate": {
            "type": "number",
            "description": "Fraction of all days with any intake logged."
          },
          "onTargetDays": {
            "type": "integer"
          },
          "totalLoggedDays": {
            "type": "integer"
          },
          "totalDays": {
            "type": "integer"
          }
        }
      },
      "Projection": {
        "type": "object",
        "required": [
          "targetWeight",
          "currentTrend",
          "weeklyRate"
        ],
        "properties": {
          "targetWeight": {
            "type": "number"
          },
          "currentTrend": {
            "type": "number"
          },
          "weeklyRate": {
            "type": "number",
            "description": "Recent slope in kg/week."
          },
          "etaWeeks": {
            "type": "number",
            "nullable": true,
            "description": "Weeks to target (null if diverging)."
          },
          "projectedDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Estimated arrival date."
          }
        }
      },
      "Insight": {
        "type": "object",
        "required": [
          "id",
          "tone",
          "title",
          "detail"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tone": {
            "type": "string",
            "enum": [
              "positive",
              "neutral",
              "warning"
            ]
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "InsightsData": {
        "type": "object",
        "required": [
          "generatedAt",
          "mode",
          "calorieTarget",
          "weeks",
          "plateau",
          "adherence",
          "expenditure",
          "highlights"
        ],
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "calorieTarget": {
            "type": "number",
            "description": "Recommended daily kcal."
          },
          "weeks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeekSummary"
            },
            "description": "Calendar-week summaries oldest to newest."
          },
          "plateau": {
            "$ref": "#/components/schemas/PlateauStatus"
          },
          "adherence": {
            "$ref": "#/components/schemas/AdherenceSummary"
          },
          "expenditure": {
            "type": "object",
            "required": [
              "estimate",
              "confidence",
              "trendDelta"
            ],
            "properties": {
              "estimate": {
                "type": "number",
                "description": "Estimated TDEE in kcal."
              },
              "confidence": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              "trendDelta": {
                "type": "number",
                "description": "Change vs 2 weeks ago."
              }
            }
          },
          "projection": {
            "$ref": "#/components/schemas/Projection"
          },
          "highlights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Insight"
            }
          }
        }
      },
...
      "ProgressData": {
        "type": "object",
        "required": [
          "generatedAt",
          "latestMeasurement",
          "measurements",
          "prs",
          "recentLifts",
          "recentCardio",
          "stepTrend",
          "streaks"
        ],
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "latestMeasurement": {
            "$ref": "#/components/schemas/Measurement"
          },
          "measurements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Measurement"
            }
          },
          "prs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LiftPR"
            }
          },
          "recentLifts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LiftEntry"
            }
          },
          "recentCardio": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardioEntry"
            }
          },
          "stepTrend": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "date",
                "steps"
              ],
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date"
                },
                "steps": {
                  "type": "integer"
                }
              }
            }
          },
          "streaks": {
            "$ref": "#/components/schemas/StreakStats"
          }
        }
      },
      "UserGoal": {
        "type": "object",
        "required": [
          "mode",
          "targetWeight",
          "startWeight",
          "startDate"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "targetWeight": {
            "type": "number",
            "description": "Goal body weight in kg."
          },
          "startWeight": {
            "type": "number",
            "description": "Starting body weight in kg."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Goal start date."
          }
        }
      },
...
      "CoachData": {
        "type": "object",
        "required": [
          "generatedAt",
          "mode",
          "headline",
          "summary",
          "focus",
          "confidence",
          "checkIn",
          "targets",
          "recommendations",
          "talkingPoints"
        ],
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "headline": {
            "type": "string",
            "description": "One-line greeting."
          },
          "summary": {
            "type": "string",
            "description": "2–3 sentence assessment."
          },
          "focus": {
            "type": "string",
            "description": "Single biggest lever."
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "checkIn": {
            "$ref": "#/components/schemas/CoachCheckIn"
          },
          "targets": {
            "$ref": "#/components/schemas/CoachTargets"
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CoachRecommendation"
            },
            "description": "Sorted by priority."
          },
          "talkingPoints": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Reflective prompts (max 4)."
          }
        }
      },
...
    }
  }
}